+++ src/test/java/org/accounting/system/ProjectEndpointTest.java
@@ -4,7 +4,6 @@ import io.quarkus.test.common.QuarkusTestResource;
 import io.quarkus.test.common.http.TestHTTPEndpoint;
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.junit.TestProfile;
-import io.quarkus.test.junit.mockito.InjectMock;
 import io.quarkus.test.keycloak.client.KeycloakTestClient;
 import io.restassured.http.ContentType;
 import org.accounting.system.clients.ProjectClient;
@@ -28,7 +27,6 @@ import org.accounting.system.repositories.client.ClientRepository;
 import org.accounting.system.repositories.metricdefinition.MetricDefinitionRepository;
 import org.accounting.system.repositories.project.ProjectRepository;
 import org.accounting.system.repositories.provider.ProviderRepository;
-import org.accounting.system.services.ReadPredefinedTypesService;
 import org.accounting.system.services.SystemAdminService;
 import org.accounting.system.services.client.ClientService;
 import org.accounting.system.util.Utility;
@@ -40,16 +38,13 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-import org.mockito.Mockito;
 
 import javax.inject.Inject;
-import java.util.Optional;
 import java.util.Set;
 import java.util.concurrent.ExecutionException;
 
 import static io.restassured.RestAssured.given;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.mockito.ArgumentMatchers.any;
 
 
 @QuarkusTest
@@ -63,9 +58,6 @@ public class ProjectEndpointTest {
     @Inject
     @RestClient
     ProjectClient projectClient;
-
-    @InjectMock
-    ReadPredefinedTypesService readPredefinedTypesService;
 
     @Inject
     @RestClient
@@ -170,14 +162,12 @@ public class ProjectEndpointTest {
     public void assignMetric(){
 
         //Registering an installation
-        Mockito.when(readPredefinedTypesService.searchForUnitType(any())).thenReturn(Optional.of("SECOND"));
-        Mockito.when(readPredefinedTypesService.searchForMetricType(any())).thenReturn(Optional.of("Aggregated"));
         var requestForMetricDefinition = new MetricDefinitionRequestDto();
 
         requestForMetricDefinition.metricName = "metric";
         requestForMetricDefinition.metricDescription = "description";
-        requestForMetricDefinition.unitType = "SECOND";
-        requestForMetricDefinition.metricType = "Aggregated";
+        requestForMetricDefinition.unitType = "TB";
+        requestForMetricDefinition.metricType = "aggregated";
 
         var metricDefinitionResponse = createMetricDefinition(requestForMetricDefinition, "admin");
 
@@ -220,14 +210,12 @@ public class ProjectEndpointTest {
     public void assignMetricAndGeneratingConflict(){
 
         //Registering an installation
-        Mockito.when(readPredefinedTypesService.searchForUnitType(any())).thenReturn(Optional.of("SECOND"));
-        Mockito.when(readPredefinedTypesService.searchForMetricType(any())).thenReturn(Optional.of("Aggregated"));
         var requestForMetricDefinition = new MetricDefinitionRequestDto();
 
         requestForMetricDefinition.metricName = "metric";
         requestForMetricDefinition.metricDescription = "description";
-        requestForMetricDefinition.unitType = "SECOND";
-        requestForMetricDefinition.metricType = "Aggregated";
+        requestForMetricDefinition.unitType = "TB";
+        requestForMetricDefinition.metricType = "aggregated";
 
         var metricDefinitionResponse = createMetricDefinition(requestForMetricDefinition, "admin");